+++ packages/nest/src/index.ts
@@ -1,4 +1,6 @@
 export * from './open-feature.module';
 export * from './feature.decorator';
+export * from './evaluation-context-interceptor';
+export * from './context-factory';
 // re-export the server-sdk so consumers can access that API from the nestjs-sdk
 export * from '@openfeature/server-sdk';